requirements.txt
crewai[google-genai]
crewai-tools
litellm
python-dotenv
requests
